+++ precompiles/staking/query_test.go
@@ -222,7 +222,7 @@ func (s *PrecompileTestSuite) TestValidator() {
 
 	testCases := []struct {
 		name        string
-		malleate    func(operatorAddress string) []interface{}
+		malleate    func(operatorAddress common.Address) []interface{}
 		postCheck   func(bz []byte)
 		gas         uint64
 		expErr      bool
@@ -230,7 +230,7 @@ func (s *PrecompileTestSuite) TestValidator() {
 	}{
 		{
 			"fail - empty input args",
-			func(operatorAddress string) []interface{} {
+			func(operatorAddress common.Address) []interface{} {
 				return []interface{}{}
 			},
 			func(_ []byte) {},
@@ -240,7 +240,7 @@ func (s *PrecompileTestSuite) TestValidator() {
 		},
 		{
 			"success",
-			func(operatorAddress string) []interface{} {
+			func(operatorAddress common.Address) []interface{} {
 				return []interface{}{
 					operatorAddress,
 				}
@@ -249,7 +249,11 @@ func (s *PrecompileTestSuite) TestValidator() {
 				var valOut staking.ValidatorOutput
 				err := s.precompile.UnpackIntoInterface(&valOut, staking.ValidatorMethod, data)
 				s.Require().NoError(err, "failed to unpack output")
-				s.Require().Equal(valOut.Validator.OperatorAddress, s.validators[0].OperatorAddress)
+
+				operatorAddress, err := sdk.ValAddressFromBech32(s.validators[0].OperatorAddress)
+				s.Require().NoError(err)
+
+				s.Require().Equal(common.HexToAddress(valOut.Validator.OperatorAddress), common.BytesToAddress(operatorAddress.Bytes()))
 			},
 			100000,
 			false,
@@ -257,11 +261,11 @@ func (s *PrecompileTestSuite) TestValidator() {
 		},
 		{
 			name: "success - empty validator",
-			malleate: func(operatorAddress string) []interface{} {
+			malleate: func(_ common.Address) []interface{} {
 				newAddr, _ := testutiltx.NewAccAddressAndKey()
 				newValAddr := sdk.ValAddress(newAddr)
 				return []interface{}{
-					newValAddr.String(),
+					common.BytesToAddress(newValAddr.Bytes()),
 				}
 			},
 			postCheck: func(data []byte) {
@@ -280,7 +284,10 @@ func (s *PrecompileTestSuite) TestValidator() {
 			s.SetupTest() // reset
 			contract := vm.NewContract(vm.AccountRef(s.address), s.precompile, big.NewInt(0), tc.gas)
 
-			bz, err := s.precompile.Validator(s.ctx, &method, contract, tc.malleate(s.validators[0].OperatorAddress))
+			operatorAddress, err := sdk.ValAddressFromBech32(s.validators[0].OperatorAddress)
+			s.Require().NoError(err)
+
+			bz, err := s.precompile.Validator(s.ctx, &method, contract, tc.malleate(common.BytesToAddress(operatorAddress.Bytes())))
 
 			if tc.expErr {
 				s.Require().Error(err)